Refactor/simplify file structur (#3831)
* moving stuff around

* Getting working again

* chore: add license

* style: format

* poke

* forgot a file

* Try on CI

* lint

* Fixed path

---------
