+++ libs/dh/market-participant/grid-areas/README.md
