--- libs/dh/market-participant/grid-areas/domain/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "dh-market-participant-grid-areas-domain",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/dh/market-participant/grid-areas/domain/src",
-  "prefix": "dh",
-  "tags": ["product:dh", "domain:market-participant", "type:domain"]
-}
