+++ libs/dh/market-participant/grid-areas/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "dh-market-participant-grid-areas-overview",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/dh/market-participant/grid-areas/overview/src",
+  "name": "dh-market-participant-grid-areas",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/dh/market-participant/grid-areas/src",
   "prefix": "dh",
   "projectType": "library",
   "targets": {
@@ -9,7 +9,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/dh/market-participant/grid-areas/overview/jest.config.ts"
+        "jestConfig": "libs/dh/market-participant/grid-areas/jest.config.ts"
       }
     },
     "lint": {
