--- libs/dh/market-participant/grid-areas/shell/README.md
@@ -1,1 +1,0 @@
-# Market Participant Grid Areas Shell
