--- libs/dh/market-participant/grid-areas/shell/project.json
@@ -1,21 +1,0 @@
-{
-  "name": "dh-market-participant-grid-areas-shell",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/dh/market-participant/grid-areas/shell/src",
-  "prefix": "dh",
-  "projectType": "library",
-  "targets": {
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/dh/market-participant/grid-areas/shell/jest.config.ts"
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"]
-    }
-  },
-  "tags": ["product:dh", "domain:market-participant", "type:shell"]
-}
