--- libs/dh/market-participant/grid-areas/shell/src/lib/dh-grid-areas-shell.routes.ts
@@ -1,32 +1,0 @@
-//#region License
-/**
- * @license
- * Copyright 2020 Energinet DataHub A/S
- *
- * Licensed under the Apache License, Version 2.0 (the "License2");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-//#endregion
-import { Routes } from '@angular/router';
-
-import { DhGridAreasShellComponent } from './dh-grid-areas-shell.component';
-
-export const dhGridAreasShellRoutes: Routes = [
-  {
-    path: '',
-    pathMatch: 'full',
-    component: DhGridAreasShellComponent,
-    data: {
-      titleTranslationKey: 'marketParticipant.gridAreas.topBarTitle',
-    },
-  },
-];
