+++ libs/dh/market-participant/grid-areas/src/graphql/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../../../../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"]
+}
