+++ libs/dh/market-participant/grid-areas/src/graphql/query/get-grid-area-overview.graphql
