+++ libs/dh/market-participant/grid-areas/src/status-badge.component.ts
