+++ libs/dh/market-participant/grid-areas/src/table.component.html
