+++ libs/dh/market-participant/grid-areas/tsconfig.lib.json
@@ -7,11 +7,6 @@
     "inlineSources": true,
     "types": []
   },
-  "exclude": [
-    "src/**/*.spec.ts",
-    "src/test-setup.ts",
-    "jest.config.ts",
-    "src/**/*.test.ts"
-  ],
+  "exclude": ["src/**/*.spec.ts", "src/test-setup.ts", "jest.config.ts", "src/**/*.test.ts"],
   "include": ["src/**/*.ts"]
 }
